Add a preprocessing capability when reading HDF5 files
Add a class with a 'preprocess(A: [])' method to the HDF5_HL module.  This
class can be inherited with the method overridden and then passed to the
various Chapel HDF5 reading functions to preprocess the input before it is
returned.

Added two example classes that inherit from the new interface class and provide
working 'preprocess()' methods.  One class takes and integer and adds that
to every element in the array.  The other class takes a string that is the
text of a script that reads from stdin and writes to stdout for each element.
It writes the script to a file, then spawns a process calling it.  It passes
the array through it using its stdin and stdout pipes.

Added new copies of existing tests that use the above two example classes.