Split the atomic comm diagnostics structs/functions into chpl-comm-diags.h
We don't want comm diagnostic implementation details exposed in the "public"
chpl-comm.h, so move them into a new chpl-comm-diags.h This also switches to
using a macro to define all the members of the comm diagnostics struct. This
cleans up the implementation by allowing us to define the members in a macro
and then just declaring ops on the macro. This also makes adding additional
fields in the future easier and prevents the atomic and non-atomic structs from
getting out of sync.