Add a .future for issue #10122
This future has an otherwise unused class as an argument to a function with a
forall loop.  It uses the class in the forall loop if it is not nil.

It hits an internal compiler error due to the reference type for the class
being missing while handling shadow variables.