Add a flag to disable the anonymous range optimization
Add a flag to disable the anonymous range optimization and turn it off under
--baseline. The anonymous range optimization was added in #1021 and #3154. It
tries to optimize iteration over anonymous ranges so that something like:

```chpl
for i in 1..10 do writeln(i);
```

doesn't actually create a temporary range just for iteration. Instead we omit
the range creation and directly iterate between the bounds of the range.
Previously, this optimization was always on and there was no way to disable it.
In hindsight there always should have been a flag to disable this optimization,
but this has only recently occurred to us.

The original implementation exposed existing leaks (details in #3195) that we
suppressed for a number of tests. Turning this off under baseline allows us to
remove the suppressions and will quiet similar failures that we're currently
seeing under --baseline testing for arrays/ferguson/pushback-no-leak and
distributions/robust/arithmetic/reindexing/simple-alias-mem-use