Add methods array.front() and array.back()
These methods return the first or last element in a 1-D rectangular array.
It is a compiler error to call them on other array types.  It is a runtime
error to call them on an empty array when bounds checking is enabled.

Added tests for these errors and a working case.