Get LLVM smoke testing back on track after disabling enum->int casts
In PR #10114, I stopped giving each enum an integer value since we're
moving towards supporting abstract enums that don't have underlying
integer values.  However, this broke LLVM code generation which seems
to generate enum values as integers (h/t to DavidK for sleuthing that
out for me).  This PR takes a stab at getting things running again,
though it may be inconsistent in doing so.

Specifically, what I did was preserve the old code (which uses the
enums' initializer values when they're available) while having cases
that don't define initializers simply define the integer values to be
the enum's order.  This seems to get testing working again, but seems
as though it could cause problems if/when there are conflicts between
a user's enum values and the ordinal values.  I _suspect_ that the
better thing to do would be to always generate enums as ordinal values
in the LLVM back-end and that this wouldn't break things now that we
generate our own enum<->int cast functions (rather than relying on the
back-end's); but this version has passed testing, so I'm going to give
it a try to get the smoke tests working again.  Then dive further into
the right way to do this.