Check for self-definitions
This PR adds a check for self-definitions like `var a: int = a;` .

While there, simplify some adjacent code.

DETAILS

The check I added works specifically on ASTs like `move(a,a)`.

I first tried a "principled approach" by making theDefinedSymbol()
return the Symbol for `a` when encountering `move(a,a)`.
Cf. currently it does so when encountering the SymExpr that is
the first arg of `move(a,a)`.
That attempt did not work for the following reason.
Consider a legitimate `move(a,b)`.  The SymExpr for `a` is encountered
before `a` is considered "defined" upon encountering the `move`. So
the SymExpr for `a` is flagged as an erroneous "use before def".