Future showing that privatized field access causes communication
This test shows that if we have a distributed array field that's
privatized (like a field that's a block array) and we only access
local elements of it, we still generate communication back to the
object's originating locale in order to access the field.  This
seems somewhere between very unfortunate and untenable.