Fix how we detect if re2 was built with valgrind support
This fixes and cleans up our detection of whether re2 was built with valgrind.
Previously start_test relied `CHPL_RE2_VALGRIND_SUPPORT` (a build time var)
being set at start_test time. This was lame because it could be wrong, and we
don't propagate CHPL_RE2_VALGRIND_SUPPORT in paratest. Further, the checks in
start_test for whether re2 was being used were really wrong. It just checked if
CHPL_REGEXP was in the environment, which meant we preventing valgrind testing
from working when CHPL_REGEXP=none.

This corrects those issues and several others in the code that checks for
valgrind compatibility. Now we:
 - Create a sentinel file in the install dir when re2 was built with valgrind
   support and use that instead of relying on CHPL_RE2_VALGRIND_SUPPORT being
   set at build and test time.
 - Only check for re2 valgrind support if re2 is actually being used!
 - Have paratest check for re2 valgrind support as well
 - Unify valgrind error messages between start_test/paratest
 - Have start_test check the target CHPL_MEM instead of host
 - Fix python style/indentation in start_test checks