Remove an enum OOB --fast suppression
In the new implementation of int->enum casts, we no longer disable OOB
checks on enums with --no-bounds-checks because it's not really saving
time compared to in-bounds cases.  If/when the implementation of enums
changes (e.g., to store orders rather than using C enums) and/or we
change/optimize the implementation of casts for common cases, this may
change again, but for now, removing the suppressif seems like the
right thing to do.