Fix error from testReleaseHelp: "./chpl-make-cpu_count: Command not found"
The fix is actually made in testRelease. This error in testReleaseHelp occurs
because testRelease copied testReleaseHelp out of its folder in the git space
before running it. With this change, testRelease just runs testReleaseHelp
directly, without copying it first.

testReleaseHelp expects to find chpl-make-cpu-count in the same directory as
itself.  If not, it will throw this error. When that happens, several
"make -j$num_proc " commands are run with "$num_proc" empty.
"make -j<nothing>" means, run as many processes in parallel as you can.
I think this has overloaded some build slaves, causing builds to crash
intermittently. However, there's no direct evidence of that.