IO.readline(): Don't return false if the last byte read happend to be \0
The bug came from using got_copy, a copy of the last byte read, as a
boolean.  I think it was being used as a test of "Did we read anything
at all?".  But I think the other tests cover all the conditions of
interest.  Remove the got_copy variable, and thereby the bug.