Convert "too many caller" halts in Barriers.chpl to boundsCheckHalts
Boundschecking is typically used for array out of bounds, but here I'm slightly
abusing the definition for use with "too many callers" to a barrier. The number
of callers is beyond the bounds specified, but it still feels a little weird to
use boundsCheckHalts here. Maybe these should just be asserts?