Convert halts in timer start()/stop() to warnings
Calling start() on a started timer or stop() on a stopped timer certainly
shouldn't be a halt, but it likely indicates a programmer error. We probably
want a richer warning interface that allows users to turn warnings on/off, but
this is at least better than halting. This also update sleep() to use the
official `warning()` wrapper which adds helpful line numbers.