Allow well written cases to pass
Creates a new check, "hasGenericArrayReturn", which determines if an array
return is not fully declared (assumed domain, generic element type, both).  If
so, any values will just get moved into the return variable and the retExprType
will be changed to dtUnknown to avoid issues.

TODO: insert checks when either the domain or the element type is specified as
part of the return value, to ensure that we don't return something we weren't
supposed to.

With this change, the original future passes.  I haven't checked the others yet.