Shrink the default stack by up to 3 system pages
Traditionally we defaulted to 8 MiB for stack task space. Qthreads also
allocated space on top of the stack for a `struct qthread_runtime_data_s`. We
default to 16 MiB huge pages on XCs, so this meant for 2 stacks we ended up
using 2 huge pages, where 1 was mostly wasted on `qthread_runtime_data_s`

Here we shrink the stack by a system page for the `qthread_runtime_data_s` and
if guard pages are enabled we shrink by another 2 pages so that everything
needed for the "stack" can fit within the 8 MiB envelope.