Add UTF-8 support for the ctype methods (#10213)
Add UTF-8 support for the ctype methods (#10213)

[reviewed by @mppf ]

This change adds UTF-8 support for the ctype methods, isAlpha() et al.

This change does not attempt to go much beyond the ctype methods. For example, split() will need to be addressed separately.

Tested in UTF-8 and non-UTF-8 environments. start_test is modified to test in a UTF-8 environment now. To avoid affecting sorting, LC_COLLATE is set to C.