Fix bigint to integral/real casts
For casts to int/uint/real we used to cast to int(64)/uint(64)/real(64) instead
of casting to the type the user specified. This corrects that issue, so we now
cast to the type requested. I also removed calls to safeCast() since these are
regular (non-safe) cast operations.

This also changes the behavior of bigint to uint casts to be the absolute value
truncated to the resultant type instead of halting for negative values. This
matches what the underlying GMP cast does, and I believe is reasonable
behavior.