Propagate environment variables to ibv-launched programs (#10293)
Propagate environment variables to ibv-launched programs (#10293)

[merged to quiet testing noise, post-commit review requested from @gbtitus and @ronawho ]

The gasnetrun_ibv launcher is written in Perl, which mangles binary strings when its UTF-8 support is turned on. Therefore, GASNet sets all the relevant character-set-support environment variables to a state that prevents UTF-8 support in the launcher. Unfortunately, this also changes the environment for the Chapel program being launched. The launcher's -E flag to propagate environment variables does not prevent this problem.

This change replicates the pre-launcher character set environment in the command line to the launcher so that the Chapel executable can support UTF-8.