Handle forall expressions over zippered arrays
In the following code:

  var C = [ab in zip(A,B)] ab;

when A and B are arrays, `ab` is a pair of references to A/B's elements.
The compiler used to try to make `C` an array of such pairs and crash.
This change makes `C` an arrays of pairs of non-references, as desired.

This issue was presented in an earlier edition of #10282.