Optimize enum->int casts for concrete types
In removing enum->int casts for abstract enums in PR #10114, I rewrote
casts for other cases to use a select statement, in part for
simplicity, and in part to generate errors for the semi-concrete case.
This turned out to have a major performance impact for fasta which
relies on a lot of enum->int casts.

Here, I'm optimizing the enum->int cast in the case where an enum is
concrete by having it fall back to the old primitive-based
implementation that used to be implemented in ChapelBase.chpl before
PR #10114.  The semi-concrete case is still handled using a switch
statement.

With more effort, other cases could also be similarlt optimized, but I
think the better use of that effort would be to implement issue
all these functions as lookups into C arrays.  So for now, I'm
stopping here in hopes that it restores the performance we expect for
benchmarks we're currently invested in and can prioritize #10307 if
other cases are required.