Improve Bessel function portability and convert halts to halt wrappers
Previously, we skipped the Bessel tests on Darwin because the 32-bit routines
weren't available. It seems pretty bad to have something in the "standard"
library that didn't work on a platform as popular as macs, so this just adds a
portability layer that will use the 32-bit routines if available, and will fall
back to the 64-bit routines with casts otherwise.

This also converts bounds checks halts to HaltWrappers.boundsCheckHalt()