Preserve promoted- and forall-expressions' shapes (#9994)
Preserve promoted- and forall-expressions' shapes

This PR adds a `_shape_` field to _iteratorRecord created for promoted expressions
and forall-expressions and uses it when creating an array out of it.

At the moment this is not a user-facing feature. That is discussed in #10359 .

Addresses a good part of #5958 .
Future work: should a range have a shape? how? See #7777.

### How does it work?

Consider a promoted + on two arrays:

    var A = B + C;

During resolution, this is converted to:

    iter +(arg1,arg2) {
      forall (a,b) in zip(arg1,arg2) do
	yield a+b;
    }

    var temp = +(B,C);
    A = chpl_initCopy(temp);

where chpl_initCopy() is one the one in ChapelArray.chpl
that takes an _iteratorRecord and creates a 1-D array
out of the things that the corresponding iterator yields.

This PR makes resolution extend the `_iteratorRecord` type of the `temp`
with a `_shape_` field and add the following statement:

    var temp = +(B,C);
    temp._shape_ = B.domain._instance;  <---------
    A = chpl_initCopy(temp);

This is done using the newly-added PRIM_ITERATOR_RECORD_SET_SHAPE
to pass information from one part of resolution to another.

chpl_initCopy() now checks whether the _iteratorRecord argument
has a meaningful shape. If so, it creates an array over that shape
instead of the 1-D array.

While ideally we would like to represent loops and loop expressions
using dedicated AST nodes, this PR achieves shape preservation
for promoted expressions with only small compiler modifications.


What's in a `_shape_` ?
-----------------------

`_shape_` keeps a pointer to a BaseDom-derived class, instead
of a ref to a `_domain`. It is simpler because the alternative
requires a ref field, which it is not obvious would work. On the
downside, we need to create a `_domain` and set its `_unownend`
to true. We can revisit this when desired.

Looks like the lifetime of the stored class instance should be OK.
That's because the promoted- or forall-expression's lifetime
should be nested in the lifetime of the source expression(s)
that it is computed from.


Test Changes
------------

* Un-futurized + gave a better name + moved to a better place -- this future requesting to preserve shapes:

    `arrays/tmacd/promotionLosesArrayShape`
    -->
    `arrays/shapes/tmacd-promotion`

* Updated .good to shape-preserved output:

    `distributions/robust/arithmetic/basics/test_promotion3`
    `studies/matrixlib/test_overloading`

* "While there": looks like this test now produced the desired output now. Un-futurize it:

    `arrays/bradc/sliceViaSingleton`

* Switch to IMO-more-elegant forall-expr-based style, for fun:

    `distributions/bharshbarg/stencil/reshape`

* Futurized this test as discussed with Michael and Brad; renamed to match a very similar variant:

    `users/vass/km/array-of-records-crash-1`
    -->
    `users/vass/km/array-of-records-crash-1.inferred-type`

* Pleasantly, remove ugly-looking adjustment that were needed due to shape divergence:

    `types/atomic/sungeun/atomic_vars`


Miscellaneous
-------------

* Moved	resolveExpr() to passes.h to avoid increased header dependence on resolution.h or iterator.h.

* While there, improved the debugging function debugGetTheIteratorFn().

Development history: 62c6da0249..ac271d3c70

r: @mppf