Do not introduce ref temps (#10010)
Do not introduce ref temps

Avoid introducing a ref temp when passing a value to a function by reference.
For example - before this change:

     def ref _ref_tmp_[1705731]:_ref(bool)[648011]
     move( _ref_tmp_[1705731] addr of( rootLocaleInitialized[94602] ) )
     call( fn =[778] _ref_tmp_[1705731] false[19] )

with this change:

     call( fn =[778] rootLocaleInitialized[94602] false[19] )

Ref temps are no longer introduced:
* by insertReferenceTemps() (now removed)
   - it was called from callDestructors.cpp and parallel.cpp
* when passing outer vars into flattened functions in addVarsToActuals()
* for the return-by-ref temp in ReturnByRef::transformMove()

Some ref temps are still inserted elsewhere in the compiler.
Avoiding those is left as future work.

Here are some things that needed adjustment to the new representation:

* removeDeadStringLiteral() is adjusted to the new pattern.

* In inferConst() in inferConstRefs.cpp, in the case
`FnSymbol* fn = call->resolvedFunction()` - check if the arg
is passed to init(), in addition to the exiting check for FLAG_RETARG.
The rest of the changes in inferConstRefs.cpp are code refactoring.
In particular. the functions onlyUsedForRetarg() and onlyUsedForInitializer()
each used to make its own pass over for_SymbolSymExprs. Now they are combined
into onlyUsedForInitOrRetarg(), which does a single pass for_SymbolSymExprs.

* denormalize() needs to distinguish passing by value and passing by ref
based on the formal's ref-ness, not the actual's ref-ness.

* parallel.cpp was using ref-ness of the actual to distinguish between
passing by ref and passing by default intent with ref semantics. It now
uses ref-ness of the type of the formal. Straightening it out is left
as future work.

* Adjusted ref detection in makeHeapAllocations(). Todo: #9106.
While there, factored out needHeapVars() in findBlockRefActuals().

* Adjusted ref detection when creating argument bundles for recursive
iterators in lowerIterators.cpp. Some ref temps are still inserted there.

* I removed a small piece in isDefinedAllPaths() in checkResolved.cpp.
That code was not invoked any longer in any of our tests.

* I removed createRetargTemps() in insertWideReferences.cpp.
Performance spot-check - thanks to BenH - did not show regression.
We may need to come back to this - and adjust it to the new representation -
if there are performance regressions elsewhere.

Development history: 08bee20f37..94677d6447

r: @benharsh