Suppress error for ZMQ with PrgEnv-cray
Sets -hnomessage to avoid messages about casting between "incompatible types",
because CCE is stricter about such circumstances than our normal compilation
with other compilers.

This is a temporary solution; we'd like to improve our handling of the type
c_fn_ptr (which is what is involved in the incompatible cast), but that's
going to require more effort and is not a high priority at the moment.

I checked that this change suppresses the warning with a basic ZMQ test and
PrgEnv-cray, and that test/release/examples/hello* and one of the hpcc
directories under test/release/examples also still work.