Add a function to read an HDF5 dataset into a distributed array
Add a function `hdf5ReadDistributedArray()` to read a dataset into a
distributed array.  It can currently work with Block and Cyclic distributed
arrays, but issues a compiler error for other distributions.  It should work
as written for any rectangular distribution for which the blocks yielded by
`localSubdomains()` are sequential in memory.

Each locale involved with the array reads its local portion of the dataset
into the array in parallel with the other locales.

Add tests that read block and cyclic arrays in 1D and 2D.  They also have
a case that would try to read block-cyclic if a param were set to true.