Add all #define'd literals from netcdf.h to NetCDF module
c2chapel only gives #define'd integer literals that are very simple - no
parenthesis or trailing comments, etc. So this commit goes through all of
the #define'd literals in netcdf.h and creates extern definitions for them
manually.

I also added a require statement for "-lnetcdf" so that it doesn't have to
be included on the command line.