Alter ModuleSymbol::getTopLevel* to return std::vectors instead of Vec
We've wanted to remove our Vec class for a long time, because reinventing the
wheel is not useful and a maintenance burden.  Such efforts have been on a "as
convenient" basis - I ran across these uses recently and so decided to remove
them while I was here.

This was mostly brainless, though I switched the uses in the docs pass to also
use sort instead of qsort.