Strengthen retInnerFunction.skipif
The future test:

    users/engin/retInnerFunction

#10010 caused this test to have different failure modes
under different configurations. #10431 reacted to that
by skipif-ing it under gasnet.

Before #10431, since this test was a .future and did not have
a .skipif, it was tested only under linux64 and gasnet.

With the addition of a .skipif in #10431, this test is now
tested under tons of configurations. Rule out the ones
with differing failure modes.