Add join() that takes forall expressions
Issue #9372 points out that, for people coming from Python, Chapel's
join() works in strange ways for forall expressions.  In the
discussion, I postulated that an overload which took a forall
expression could potentially be provided to address this confusion.
This PR adds such an overload (with no attempts to optimize it for
parallelism or minimizing string operation overheads).