Add a test to lock in the expected behavior with --makefile
Figuring out how to test this appropriately was a bit tricky.  I decided to
take the approach of making a sample Makefile that used the generated one to
compile and run a C test, saving the output of that into the generated output
file.

I could have explicitly checked what the generated Makefile would contain, but
since that would vary between compilation environments, I would have needed to
generate the expected output anyways and that wouldn't lock in that the set up
actually worked, just that the file looked the way I expected it to.

I added a sub_test file instead of overriding the header name (since -o modifies
the library name and the header file name is based on of that when it is not
explicitly set).