Write the generated Makefile directly in the executable location.
This is a temporary step.  We would like to make a "lib" directory for the
generated files to live so that they don't clutter up the directory, and the
Makefile should go there.  But I didn't want to update the runtime Makefile to
"move" the generated one from the temporary directory or generated code
directory into the place where the final library lives (like we do for the
library file and the header file at the moment), since we were just going to
change it.