Add a test for the atomics API
This adds pretty comprehensive test for the atomic API for all supported atomic
types. It tests all sequentially-consistent functions. There's a -nl1 test, and
a -nl2 variant that tests with and without network atomics and with updates
done from a remote locale and from the locale the atomic is declared on.

The -nl1 variant tests that:
 - the basic API (for processor and network atomics implementations) works
 - the special case code in ugni for `chpl_numNodes == 1` works

The -nl2 variant tests that:
 - remote updates to processor atomics works (implementation uses on-stmts)
 - the "normal" (numNodes != 1) code in ugni for network atomics works
 - ugni works with processor and network atomics