Convert uses of internal processor atomic types to chpl__processorAtomicType
Convert uses of things like `atomic_int64` to `chpl__processorAtomicType(int)`.
The atomic_* types are the internal names for the processor atomic
implementation, but they're going to be removed as part of #10493.

chpl__processorAtomicType() isn't user-facing either, but it's the best we have
until we implement a user-facing mechanism to force processor atomics. See
https://github.com/chapel-lang/chapel/issues/10551 for more information.