Fix timeouts for high core-count systems (#10530)
Fix timeouts for high core-count systems (#10530)

[reviewed by @ronawho ]

In some tests, high core-count systems spend most of their time in the mprotect() system call when guard pages are used in qthreads. The arm-thunderx has 96 cores per locale, and arm-thunderx2 has 64. This change turns off guard pages for two tests where this makes a big difference.

One remaining item to investigate is that on arm-thunderx2, there appears to be more than just a performance problem. Intermittently, high contention in the mprotect() system call can lead to an application hanging indefinitely, which might be an indication of a kernel or hardware bug.  This is documented in #7533 for further investigation.

Closes #10475