Avoid communication in init_elts
Previously we did `if here == dummyLocale` to check if `here` is initialized.
`dummyLocale` lives on locale 0, so this results in communication for non-0
locales (which is unnecessary comm and breaks array init inside of local blocks.)

This switches the check to look for if !rootLocaleInitialized` (where
`rootLocaleInitialized` is locale private/replicated.) 8363cdcee1 added the
`rootLocaleInitialized` for similar comm-avoidance reasons and may provide some
more context.