chpl_launchcmd.py should wait a bit for qsub output file to appear
Under certain conditions, chpl_launchcmd may throw an exception
(and thus a false test error) because it looks for the qsub output
file before that file actually exists on the file system where
chpl_launchcmd is running. Maybe because chpl_launchcmd is running
on an "elogin" machine, and the NFS or whatever file transport is
slow to deliver the file from the real login node to the elogin
machine. If chpl_launchcmd allowed a little extra time for the
qsub output file to show up, everything might turn out Okay.

This change implements such a scheme with a max extra waiting time
of 30 seconds. If any extra waiting time occurred (ie, the output
file was not available immediately), a log.warn message is issued.

The extra waiting time (if any) is subtracted from the execution
time reported by sub_test as
"launchcmd reports elapsed execution time"