Move require statements into Chapel functions
This makes the requirement of the BLAS headers dependent on whether or
not the user calls any of the BLAS functions.

One potential downside of this approach is that the user needs to
explicitly require the BLAS headers if using C_BLAS module. Fortunately,
users should rarely be interacting with this interface.