Remove added "name.c:" output from these tests
When compiling with multiple files, sometimes PrgEnvs will output the C file
used.  We were avoiding this in our testing system with an update for sub_test
from 2015, but these tests avoided that guard (because I was compiling through
a makefile instead of the testing system).  To avoid this issue, use sed to
remove the output when it is there.