Adjust skyline.bad
This should fix skyline.bad mismatch in nightly valgrind testing.

This test started failing in a different way.
I am updating .bad and .prediff to accomodate.

The new .bad and .prediff are gcc-specific.
This works because we test this test in the nightlies
only under valgrind.

This test is on my short list to investigate already.