Minor cleanup+reorg of ForallStmt-related code
* Importantly, move implementForallIntentsNew() to go AFTER
resolveParallelIteratorAndIdxVar(). This sets me up for
merging implementForallIntents1New() and setupShadowVariables(),
which do different aspects of the same task.

* Get rid of implementForallIntentsNew(), instead call
implementForallIntents1New() (with the `1`) and checkForNonIterator()
directly.

* Note that placing checkForNonIterator() requires some care.
(This was the case before as well.) It needs to happen after
the iterator is resolved, and before its yield time is queried.

* Remove the unused field reduceGlobalOp.

* list_ast() now produces prettier output for ForallStmts.

* Move ForallStmt- and UseStmt-specific code from list_ast()
into helper functions.

* Relocate print_indent() and print_on_its_own_line() towards
the top of view.cpp as they are used by other helpers and
have more general applicability.