Force GASNet to build the specified substrate
Previously, we just let GASNet auto-detect the substrate/conduit. This meant
that a user could explicitly build with CHPL_COMM_SUBSTRATE=mpi and even if mpi
wasn't available on the system we wouldn't error, which led to much more
cryptic build or even runtime issue later on.

Now we throw `--enable-$CHPL_COMM_SUBSTRATE` to the GASNet build and if GASNet
can't build the specified substrate it will stop with a clear error message.