Add a compilerErrorCont module function
compilerError calls USR_FATAL(), which immediately stops compilation. This
makes it annoying to test the atomic error messages, but also seems like it's
probably a pain for users since they have to work through a single compiler
error at a time.

Add a compilerErrorCont that calls USR_FATAL_CONT -- I wonder if this should be
the default for compilerError?