Add degenerate chpl_deinitModules in minimal modules
We don't deinit modules in minimal modules compilation.  There was an empty
version of the addModule function already, so add an empty version of
chpl_deinitModules to complement it (and avoid having an undefined reference
error when compiling in that mode, since we need to match the symbol at
compilation time).