Prototype support for "always RVF" types
This adds a pragma "always RVF" that can be attached to a type in
order to request that the compiler always RVF objects of that type.
This is currently intended as an experimental feature, not for typical
end-users.  Over time, I think we need to think about whether this is
something we'd want to expose in the language, and if so, how to do so
(where one way might be to introduce a notion of 'intents' for
on-clauses and to permit type authors to override the default intent
of `ref` used for all types presently)..

Added a simple test to prove to myself that this feature is working.

Note that, as it stands, use of this feature can result in surprising
differences between --local and --no-local compilations since
r.v.f. doesn't have a role in --local compilations.  The difference
between the .good and .comm-none.good files in my test demonstrates
this difference.  If we were to promote this to a full language
concept, it might suggest that we should always r.v.f. / copy-in such
types across on-clauses to keep behavior as similar as possible
between the two modes.

OTOH, for cases where this feature arguably _ought_ to be used (ones
in which additional copies of the record and modifications to those
copies won't affect the object's semantics -- as with a record
wrapping an invariant class object where the class stores everything
important about the semantic state -- maybe it doesn't matter whether
the copy is made or not (and avoiding the copy would be preferable).
Put another way: The record in my example is not one which should ever
have this pragma attached to it, so maybe the variable output is
completely reasonable and as expected.

I think there are interesting overlaps between this feature and
serialization; and in a recent discussion, BHarsh and I have been
discussing whether this feature + serialization would be a way to have
distributed slices be privatized lazily rather than aggressively as
they are today, optimizing away the overhead of creating them.